# doc/swagger -event.yaml
swagger: '2.0'
info:
  description: >-
    信息接报、事件管理、应急处置、处置跟踪、总结评估等.
  version: 1.0.0
  title: event事件处置
host: 127.0.0.1:8000
basePath: /emp-event/v1
...
schemes:
  - http
paths:
...
 /cmd/eventLog:
   post:
     tags:
       - '【浙能需求】应急处置-应急日志cmdEventLog'
     summary: '新增事件应急日志'
     operationId: addEventLog
     parameters:
       - name: eventLog
         in: body
         required: true
         schema:
           $ref: '#/definitions/EventLog'
     responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ReturnResult'
...
 /cmd/schemes:
    post:
       tags:
         - 应急处置-方案管理cmdScheme
       summary: '启动预案，生成处置方案基本信息，处置方案名称默认为‘eventName+处置方案’'
       operationId: addEcScheme
       parameters:
         - name: ecScheme
           in: body
           required: true
           schema:
             type: object
             properties:
               eventId:
                 type: string
                 description: 事件ID
               eventName:
                 type: string
                 description: 事件名称
       responses:
         200:
           description: OK
           schema:
             type: object
             allOf:
                - $ref: '#/definitions/ReturnResult'
             properties:
                data:
                    $ref: '#/definitions/Scheme'
    get:
      tags:
          - 应急处置-方案管理cmdScheme
      summary: '根据事件ID获取事件处置方案基本信息'
      operationId: findSchemeVoByEventId
      parameters:
        - name: eventId
          in: query
          description: 事件ID
          type: string
      responses:
        200:
          description: OK
          schema:
            type: object
            allOf:
              - $ref: '#/definitions/ReturnResult'
            properties:
              data:
                $ref: '#/definitions/Scheme'
...
 /cmd/experts:
    post:
      tags:
        - 应急处置-应急专家调度cmdExpert
      summary: '新增关联应急专家'
      operationId: addEcExperts
      parameters:
        - name: ecExperts
          in: body
          required: true
          schema:
            type: object
            properties:
              schemeId:
                type: string
                description: 处置方案ID
              expertIds:
                type: array
                description: 关联的专家ID,数组
                items:
                  type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ReturnResult'
...
definitions:
...
 EventLog:
   type: object
   properties:
     id:
       type: string
     eventId:
       type: string
     eventName:
       type: string
     logContent:
       type: string
     operator:
       type: string
     operatorTime:
       type: string
       format: 'date-time'
 Scheme:
    type: object
    properties:
      id:
        type: string
      schemeName:
        type: string
        description: 可编辑
      eventId:
        type: string
...
 ReturnResult:
    type: object
    properties:
      code:
        type: integer
        description: >-
          2000  操作成功
          4000 操作失败
          4001 验证码出错
          4002 用户名或密码错误
          5000 后台逻辑错误
          5001 令牌失效或过期
          5009 未知错误
      msg:
        type: string

############################ parameters#####################
...